Remove explicit check for source type
Instead of checking the options hash for source type, we should be able
to now rely on the `type` of the reflection to see if that is set. If
that is set then its a polymorphic and we should just follow the steps
that rails does with when applying scope for this.