setup tables and relationship for polymorphic through a has_one
This adds models and relationships that will represent a polymorphic
relationship that will have a through that goes `has_one` model. We ran
into this issue while working converting a relationship in
github/github.

We are using the same(ish) models of that we use in dotcom. This will
make it easier to mimic the behavior.