Order records in memory
If a previous reflection has an order, pass that order to a new class
that will order the records in memory.

When you join 2 tables where one has an order that order will be applied
to the returned records, even if that order isn't explcitly on that
table. For example if you have repos that are pinned to a profile and
set in a certain order the profile pins contain the position column for
the order and are applied to repos. However once the query is split in 2
the profile pins can't pass the order column to repos, the repos have no
position column.

To fix this we can take the order and apply it in memory using Ruby.
This is not super efficient for large sets so we added an additional
warning if there are more than 5k records. We also make sure that the
records are uniqued when ordering. mysql would return all 3 records with
a join, but since it's difficult to return that in all cases, especially
without ordering, and expensive to order in memory, we should always
return a distinct dataset.

We had considered using the FIELD order query to order database records
by value in the query itself, but that is different for every database
and is also inefficient. This is better in that we don't need to
determine how each type of database does the ordering.

Co-authored-by: Aaron Patterson <aaron.patterson@gmail.com>