Ensure last applied order wins
We don't want to apply the order from a previous scope if the returned
records have their own order. For example, pinned ships have an order on
pin position, but if ships have their own order applied that order
should win.

This changes the code to check if the order values are empty and if the
records are ordered. That way if the order values aren't empty we can
fall through to `scope` which will let mysql add the order rather than
Ruby - and make sure that order is correct, because last order should
win.